+++ README.md
@@ -15,3 +15,24 @@ Dix is a work in progress, and the following features are planned for implementa
 - Word count and line count display for effective content management
 
 
+
+## Run:
+
+### Prerequisit:
+
+you should have cmake installed in your system
+
+**for linux**:
+
+`sudo apt-get install cmake`
+
+then run: 
+
+```shell
+chmod +x ./run.sh
+./run.sh
+```
+
+
+MIT License
+Copyright (c) 2023 duckduckcodes
