+++ run.sh
@@ -1,4 +1,16 @@
 #!/bin/bash
 
-gcc -o main main.c
-./main
+# Check if build directory exists, if not, create it
+if [ ! -d "build" ]; then
+  mkdir build
+fi
+
+# Navigate to the build directory
+cd build
+
+# Run cmake and make
+cmake ..
+make
+
+# Execute the built program
+./dix
