+++ src/editor.c
@@ -1,63 +1,4 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <termios.h>
-#include <string.h>
-#include <stdbool.h>
-#include <unistd.h>
-
-#define CHECK_MEMORY(ptr)                                                               \
-    do                                                                                  \
-    {                                                                                   \
-        if (!(ptr))                                                                     \
-        {                                                                               \
-            fprintf(stderr, "Memory allocation failed at %s:%d\n", __FILE__, __LINE__); \
-            exit(EXIT_FAILURE);                                                         \
-        }                                                                               \
-    } while (0)
-
-#define SAFE_MALLOC(ptr, size) \
-    do                         \
-    {                          \
-        ptr = malloc(size);    \
-        CHECK_MEMORY(ptr);     \
-    } while (0)
-
-#define SAFE_REALLOC(ptr, size)              \
-    ({                                       \
-        void *temp_ptr = realloc(ptr, size); \
-        CHECK_MEMORY(temp_ptr);              \
-        temp_ptr;                            \
-    })
-
-enum Keypress
-{
-    ENTER = 10,
-    ESCAPE = 27,
-    BACKSPACE = 127,
-    TAB = 9,
-    CTRL_C = 3,
-    CTRL_S = 19,
-    CTRL_V = 22,
-    CTRL_X = 24,
-};
-
-typedef struct
-{
-    char *lineContent;
-    int position;
-    int line;
-
-} Line;
-
-typedef struct
-{
-    int cursor_x, cursor_y;
-    Line **lines;
-    char *filename;
-    bool saved;
-    int lineCount;
-    int characterCount;
-} EditorState;
+#include "editor.h"
 
 static struct termios orig_termios; /* In order to restore at exit. */
 
@@ -103,6 +44,7 @@ void updateScreen(EditorState *e)
         }
     }
 }
+
 
 void insertChar(char addedChar, EditorState *e)
 {
@@ -208,39 +150,4 @@ void freeEditor(EditorState *e)
         free(e->lines[i]);
     }
     free(e->lines);
-}
-
-int main()
-{
-    enableRawMode(STDIN_FILENO);
-    system("clear");
-
-    EditorState editor = {
-        .cursor_x = 0,
-        .cursor_y = 0,
-        .filename = "empty",
-        .lines = NULL,
-        .lineCount = 0,
-        .characterCount = 0,
-        .saved = false};
-
-    initiateNewLine(&editor);
-
-    char c;
-    while (1)
-    {
-        if (read(STDIN_FILENO, &c, 1) == 1)
-        {
-            if (c == CTRL_X)
-            {
-                break;
-            }
-            else
-            {
-                insertChar(c, &editor);
-            }
-        }
-    }
-    freeEditor(&editor);
-    return 0;
 }
